+++ src/main/java/com/satyamevjayate/api/entity/CrimeType.java
@@ -30,7 +30,7 @@ public class CrimeType {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name="CrimeTypeID")
-	private BigInteger crimeTypeId;
+	private Long crimeTypeId;
 	
 	@Column(name="CrimeType")
 	private String crimeType;
