+++ src/main/java/com/satyamevjayate/api/repo/CrimeSuspectRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface CrimeSuspectRepository extends JpaRepository<CrimeSuspect, BigInteger> {
+public interface CrimeSuspectRepository extends JpaRepository<CrimeSuspect, Long> {
 }
