+++ src/main/java/com/satyamevjayate/api/repo/CrimeVictimDocumentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface CrimeVictimDocumentRepository extends JpaRepository<CrimeVictimDocument, BigInteger> {
+public interface CrimeVictimDocumentRepository extends JpaRepository<CrimeVictimDocument, Long> {
 }
