+++ src/main/java/com/satyamevjayate/api/repo/LawActRepository.java
@@ -3,6 +3,6 @@ import java.math.BigInteger;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.satyamevjayate.api.entity.LawAct;
 
-public interface LawActRepository extends JpaRepository<LawAct,BigInteger> {
+public interface LawActRepository extends JpaRepository<LawAct,Long> {
 
 }
