+++ src/main/java/com/satyamevjayate/api/repo/PersonRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface PersonRepository extends JpaRepository<Person, BigInteger> {
+public interface PersonRepository extends JpaRepository<Person, Long> {
 }
